requirements.txt
requests==2.32.3
beautifulsoup4==4.13.3
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.40
redis==5.2.1
Celery==5.5.1
psycopg2-binary==2.9.10
chromadb==1.0.8